- Implementing the logic of the Service layer methods for each entity in the HealthInsurance service. - Implementing logs in each service. -Generating corresponding updates in the dtos, repositories for correct implementation of business logic. -Creation of DuplicateInformationException to handle exceptions for information that cannot pre-exist in the database. -Creating the missing dtos: PatientInsuranceCreateDto, PatientInsuranceUpdateDto and PatientInsuranceResponseDto.

+++ HealthInsuranceService/src/main/java/com/sanatoryApp/HealthInsuranceService/dto/Request/HealthInsuranceCreateDto.java
@@ -21,7 +21,7 @@ public class HealthInsuranceCreateDto {
     @Size(min = 5,max = 50, message = "The name must be between 5 and 50 characters long.")
     private String companyName;
     @Positive(message = "Company's Code must be positive.")
-    private int companyCode;
+    private Long companyCode;
     @NotBlank
     @Pattern(regexp = "^\\+[1-9]\\d{1,14}$", message = "Invalid international phone format")
     @Schema(description = "Phone Number in international format ",example = "+5491112345678")
